DBParser: Fix NullPointerException on file without extension.

+++ products/ITER/plugins/org.csstudio.utility.dbparser/src/org/csstudio/utility/dbparser/StartupLauncher.java
@@ -65,9 +65,11 @@ public class StartupLauncher implements IStartup {
 					event.getDelta().accept(new IResourceDeltaVisitor() {
 						public boolean visit(IResourceDelta delta) {
 							IResource resource = delta.getResource();
-							if (!(resource.getType() == IResource.FILE))
+							if (resource.getType() != IResource.FILE)
 								return true;
-							if (resource.getFileExtension().toLowerCase().equals("db")) {
+							if (resource.getFileExtension() != null
+									&& resource.getFileExtension()
+											.toLowerCase().equals("db")) {
 								switch (delta.getKind()) {
 								case IResourceDelta.REMOVED:
 									DBContextValueHolder.get().removeFile((IFile) resource);
